requirements.txt
pandas
numpy
scikit-learn
xgboost
fairlearn
shap
pyyaml
openpyxl